vink_chat_bot-master/backend/vink/api/serializers.py
from rest_framework import serializers
from .models import Product


class ProductSerializer(serializers.ModelSerializer):
    text = serializers.CharField()

    class Meta:
        model = Product
        fields = ['text',]
