vink_chat_bot-master/backend/requirements.txt
Django==2.2.16
requests==2.26.0
sorl-thumbnail
Faker
djangorestframework==3.15.1
django-cors-headers==4.3.1